+++ src/Symfony/Component/HttpClient/Tests/Psr18ClientTest.php
@@ -74,4 +74,13 @@ class Psr18ClientTest extends TestCase
         $this->expectException(Psr18RequestException::class);
         $client->sendRequest($factory->createRequest('BAD.METHOD', 'http://localhost:8057'));
     }
+
+    public function test404()
+    {
+        $factory = new Psr17Factory();
+        $client = new Psr18Client(new NativeHttpClient());
+
+        $response = $client->sendRequest($factory->createRequest('GET', 'http://localhost:8057/404'));
+        $this->assertSame(404, $response->getStatusCode());
+    }
 }