+++ src/Symfony/Component/Mailer/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+4.4.0
+-----
+
+ * [BC BREAK] Transports depend on `Symfony\Contracts\EventDispatcher\EventDispatcherInterface`
+   instead of `Symfony\Component\EventDispatcher\EventDispatcherInterface`.
+
 4.3.0
 -----